# Use a reference to C memory instead of a copy for BLOBs