This changes Scan to assign a reference to a BLOBs C memory instead of creating a copy and is safe because the database/sql package creates a copy unless the destination is a sql.RawBytes. Basically, this savs a useless alloc.

This change may break programs that use *sqlite3.SQLiteRows directly instead of through the sql package.

The `sqlite_copy_bytes` build tag was added to disable this behavior and revert to the prior behavior of copying the byte slice.

```
goos: darwin
goarch: arm64
pkg: github.com/charlievieth/go-sqlite3
cpu: Apple M4 Pro
                               │   x1.txt    │               x2.txt                │
                               │   sec/op    │   sec/op     vs base                │
Suite/BenchmarkScanRawBytes-14   3.084µ ± 6%   2.658µ ± 0%  -13.81% (p=0.000 n=10)

                               │   x1.txt    │               x2.txt               │
                               │    B/op     │    B/op     vs base                │
Suite/BenchmarkScanRawBytes-14   8784.0 ± 0%   592.0 ± 0%  -93.26% (p=0.000 n=10)

                               │   x1.txt   │               x2.txt               │
                               │ allocs/op  │ allocs/op   vs base                │
Suite/BenchmarkScanRawBytes-14   26.00 ± 0%   18.00 ± 0%  -30.77% (p=0.000 n=10)
```